Removed extraneous utils
